data type fix and position added

+++ VoxTerGenEngine/include/graphics/Camera.hpp
@@ -71,7 +71,7 @@ public:
      * @param pitch Initial pitch angle in degrees
      */
     Camera(
-        glm::vec3 position = glm::vec3(10.0, 15.0, 10.0), 
+        glm::dvec3 position = glm::vec3(10.0, 15.0, 10.0), 
         glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f),
         float yaw = constants::camera::yaw, 
         float pitch = constants::camera::pitch);
